crud refactor 24-01-24 and refactor TINYINT to Integer

+++ src/mmisp/db/models/tag.py
@@ -1,5 +1,4 @@
 from sqlalchemy import Column, Integer, String
-from sqlalchemy.dialects.mysql import TINYINT
 
 from ..database import Base
 
@@ -10,14 +9,14 @@ class Tag(Base):
     id = Column(Integer, primary_key=True)
     name = Column(String(255), unique=True)
     colour = Column(String(255))
-    exportable = Column(TINYINT)
+    exportable = Column(Integer)
     org_id = Column(Integer, index=True)
     user_id = Column(Integer, index=True)
-    hide_tag = Column(TINYINT)
+    hide_tag = Column(Integer)
     numerical_value = Column(Integer, index=True)
-    is_galaxy = Column(TINYINT)
-    is_custom_galaxy = Column(TINYINT)
+    is_galaxy = Column(Integer)
+    is_custom_galaxy = Column(Integer)
     attribute_count = Column(Integer)  # new
     count = Column(Integer)  # new
-    favourite = Column(TINYINT)  # new
-    local_only = Column(TINYINT)  # new
+    favourite = Column(Integer)  # new
+    local_only = Column(Integer)  # new
